Use proper HttpException as base
