+++ src/Graviton/ExceptionBundle/Exception/RestException.php
@@ -5,6 +5,7 @@
 
 namespace Graviton\ExceptionBundle\Exception;
 
+use Symfony\Component\HttpKernel\Exception\HttpException;
 use Symfony\Component\HttpFoundation\Response;
 
 /**
@@ -14,36 +15,7 @@ use Symfony\Component\HttpFoundation\Response;
  * @license  http://opensource.org/licenses/gpl-license.php GNU Public License
  * @link     http://swisscom.ch
  */
-abstract class RestException extends \Exception
+abstract class RestException extends HttpException implements RestExceptionInterface
 {
-    /**
-     * Response object
-     *
-     * @var Response
-     */
-    private $response = false;
-
-    /**
-     * Set the response object (optional)
-     *
-     * @param Response $response Response object
-     *
-     * @return RestException $this This
-     */
-    public function setResponse(Response $response)
-    {
-        $this->response = $response;
-
-        return $this;
-    }
-
-    /**
-     * Get the response object
-     *
-     * @return Response $response Response object
-     */
-    public function getResponse()
-    {
-        return $this->response;
-    }
+    use RestExceptionTrait;
 }
