+++ src/Graviton/ExceptionBundle/Exception/SerializationException.php
@@ -24,6 +24,6 @@ final class SerializationException extends RestException
      */
     public function __construct($message = "Serialization Error", $prev = null)
     {
-        parent::__construct($message, Response::HTTP_INTERNAL_SERVER_ERROR, $prev);
+        parent::__construct(Response::HTTP_INTERNAL_SERVER_ERROR, $message, $prev);
     }
 }
